requirements.txt
openai>=1.40.0
pytest>=8.0.0
requests>=2.31.0
pyyaml>=6.0.0
allure-pytest>=2.13.2
json5>=0.9.25
